Refactored behavior import/export actions into new controller - also fixed up some jsRoutes
