+++ app/views/intro.scala.html
@@ -10,7 +10,7 @@
 
 @main("Welcome to Ellipsis", Some(teamAccess), "intro") {
   @helper.javascriptRouter("jsRoutes")(
-    routes.javascript.ApplicationController.doImportBehavior,
+    routes.javascript.BehaviorImportExportController.doImport,
     routes.javascript.ApplicationController.editBehavior
   )
   <div id="content" class="content">
