+++ conf/routes
@@ -8,10 +8,10 @@ GET         /intro                                    @controllers.ApplicationCo
 GET         /install_behaviors                        @controllers.ApplicationController.installBehaviors(teamId: Option[String] ?= None)
 GET         /new_behavior                             @controllers.ApplicationController.newBehavior(teamId: Option[String] ?= None)
 GET         /edit_behavior/:id                        @controllers.ApplicationController.editBehavior(id: String, justSaved: Option[Boolean] ?= None)
-GET         /export_behavior/:id                      @controllers.ApplicationController.exportBehavior(id: String)
-GET         /import_behavior_zip                      @controllers.ApplicationController.importBehaviorZip(teamId: Option[String] ?= None)
-POST        /import_behavior_zip                      @controllers.ApplicationController.doImportBehaviorZip
-POST        /import_behavior                          @controllers.ApplicationController.doImportBehavior
+GET         /export_behavior/:id                      @controllers.BehaviorImportExportController.export(id: String)
+GET         /import_behavior_zip                      @controllers.BehaviorImportExportController.importZip(teamId: Option[String] ?= None)
+POST        /import_behavior_zip                      @controllers.BehaviorImportExportController.doImportZip
+POST        /import_behavior                          @controllers.BehaviorImportExportController.doImport
 POST        /save_behavior                            @controllers.ApplicationController.saveBehavior
 POST        /delete_behavior                          @controllers.ApplicationController.deleteBehavior
 POST        /clone_behavior                           @controllers.ApplicationController.cloneBehavior
